+++ settings.py
@@ -1,8 +1,8 @@
 import math
 
 # game settings
-RES = WIDTH, HEIGHT = 860, 540
-# RES = WIDTH, HEIGHT = 1920, 1080
+# RES = WIDTH, HEIGHT = 860, 540
+RES = WIDTH, HEIGHT = 1920, 1080
 HALF_WIDTH = WIDTH // 2
 HALF_HEIGHT = HEIGHT // 2
 FPS = 0
